requirements.txt
Django==1.11.8
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
dwebsocket==0.5.10
paho-mqtt==1.4.0
Pillow==6.1.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
